+++ composenav/src/main/java/com/adamkobus/compose/navigation/intent/NavIntentResolvingManager.kt
@@ -2,7 +2,7 @@ package com.adamkobus.compose.navigation.intent
 
 import com.adamkobus.compose.navigation.NavIntentResolver
 import com.adamkobus.compose.navigation.action.NavAction
-import com.adamkobus.compose.navigation.destination.CurrentDestination
+import com.adamkobus.compose.navigation.destination.NavState
 
 internal class NavIntentResolvingManager {
 
@@ -12,9 +12,9 @@ internal class NavIntentResolvingManager {
         this.resolvers.addAll(resolvers)
     }
 
-    suspend fun resolve(intent: NavIntent, currentDestination: CurrentDestination): NavAction? {
+    suspend fun resolve(intent: NavIntent, navState: NavState): NavAction? {
         val history = NavIntentHistory(intent)
-        val result = getNextResult(intent, currentDestination, history)
+        val result = getNextResult(intent, navState, history)
         return if (result is ResolveResult.Action) {
             result.action
         } else {
@@ -22,15 +22,15 @@ internal class NavIntentResolvingManager {
         }
     }
 
-    private suspend fun getNextResult(intent: NavIntent, currentDestination: CurrentDestination, history: NavIntentHistory): ResolveResult {
+    private suspend fun getNextResult(intent: NavIntent, navState: NavState, history: NavIntentHistory): ResolveResult {
         resolvers.forEach {
-            val resolverResult = it.resolve(intent, currentDestination)
+            val resolverResult = it.resolve(intent, navState)
             if (resolverResult is ResolveResult.Action) {
                 return resolverResult
             }
             if (resolverResult is ResolveResult.Intent) {
                 history.addNode(resolverResult.intent)
-                val next = getNextResult(resolverResult.intent, currentDestination, history)
+                val next = getNextResult(resolverResult.intent, navState, history)
                 history.popNode(resolverResult.intent)
                 if (next is ResolveResult.Action) {
                     return next