+++ demo-common-ui/src/main/java/com/adamkobus/compose/navigation/demo/ui/nav/DogsBrowserGraph.kt
@@ -1,9 +1,9 @@
 package com.adamkobus.compose.navigation.demo.ui.nav
 
-import androidx.navigation.NavBackStackEntry
 import com.adamkobus.compose.navigation.destination.NavGraph
+import com.adamkobus.compose.navigation.destination.NavStackEntry
 import com.adamkobus.compose.navigation.destination.ScreenDestination
-import com.adamkobus.compose.navigation.ext.getInt
+import com.adamkobus.compose.navigation.destination.getInt
 
 object DogsBrowserGraph : NavGraph("dogs") {
     const val PARAM_DOG_ID = "dogId"
@@ -18,4 +18,4 @@ object DogsBrowserGraph : NavGraph("dogs") {
     val Back = popDestination()
 }
 
-fun NavBackStackEntry.dogId() = getInt(DogsBrowserGraph.PARAM_DOG_ID)
+fun NavStackEntry?.dogId() = getInt(DogsBrowserGraph.PARAM_DOG_ID)